CI: arreglar el aviso C4244 de MSVC que tumbaba toda la matriz de Windows
Con las dependencias de Python puestas, Linux paso entero y quedo al
descubierto el fallo REAL de Windows, que llevaba tapado detras del otro.

En step.hpp, la comprobacion de poblacion de TRAIN_UNIT pasa `c.emitter`
(uint16_t) a `player_pop_cap`, que toma uint8_t. gcc y clang callan; MSVC
avisa con C4244 y, con -Werror, eso no rompe una prueba: rompe la COMPILACION
de los 30 objetivos del proyecto.

Cast explicito. La conversion es segura y esta probada arriba en el mismo
fichero: el despacho rechaza toda orden con emitter >= MAX_EMITTERS, y
MAX_EMITTERS es 16 — cabe de sobra en un uint8_t. El cast no tapa el aviso, lo
resuelve.

Sin cambio de comportamiento: no sube ninguna version de algoritmo ni mueve
ningun digest. Suite 51/51.

+++ addons/chunsa_sim/core/include/chunsa/step.hpp
@@ -571,7 +571,14 @@ inline RejectReason apply_command(GameState& g, const ScheduledCommand& c) noexc
             // Sprint 1.14: el tope ya no es la constante regalada, sino lo que
             // el jugador HAYA CONSTRUIDO. POP_CAP_V1 sigue dentro, como cota
             // dura, en player_pop_cap.
-            if (g.pop_used[c.emitter] + pop_cost > player_pop_cap(g, c.emitter)) {
+            // El cast es explícito a propósito: `emitter` es uint16_t y
+            // `player_pop_cap` toma uint8_t. MSVC avisa (C4244) donde gcc y
+            // clang callan, y con -Werror eso tumba TODA la matriz de Windows.
+            // La conversión es segura y está probada arriba: el despacho
+            // rechaza cualquier orden con `emitter >= MAX_EMITTERS`, y
+            // MAX_EMITTERS es 16.
+            if (g.pop_used[c.emitter] + pop_cost
+                > player_pop_cap(g, static_cast<uint8_t>(c.emitter))) {
                 return RejectReason::ILLEGAL_STATE;
             }
